Move static FuzzySearch to its own class
